Used a different Serializer to store data, If in future much things need to saved in the state object thus rather than storing Only List of WidgetTaskModel we will store SnaptickWidgetState which for the meantime will only store widget tasks
